@@ -2,8 +2,9 @@ package main
 
 import (
 	"fmt"
-	"github.com/gbrayhan/microservices-go/config"
-	errorsController "github.com/gbrayhan/microservices-go/controllers/errors"
+	"github.com/gbrayhan/microservices-go/infrastructure/repository/config"
+	errorsController "github.com/gbrayhan/microservices-go/infrastructure/rest/controllers/errors"
+	"github.com/gbrayhan/microservices-go/infrastructure/rest/middlewares"
 	"net/http"
 	"strings"
 	"time"
@@ -13,37 +14,32 @@ import (
 	"github.com/gin-gonic/gin"
 	"github.com/spf13/viper"
 
-	"github.com/gbrayhan/microservices-go/middlewares"
-	"github.com/gbrayhan/microservices-go/routes"
+	"github.com/gbrayhan/microservices-go/infrastructure/rest/routes"
 )
 
 func main() {
 	router := gin.Default()
-	initialGinConfig(router)
-	router.Use(middlewares.GinBodyLogMiddleware)
-	router.Use(errorsController.Handler)
-	routes.ApplicationV1Router(router)
-	startServer(router)
-
-}
-
-func initialGinConfig(router *gin.Engine) {
 	router.Use(limit.MaxAllowed(200))
 	router.Use(cors.Default())
 	var err error
-	config.DB, err = config.GormOpen()
-
+	DB, err := config.GormOpen()
 	if err != nil {
-		panic(fmt.Errorf("fatal error in database file: %s \n", err))
+		_ = fmt.Errorf("fatal error in database file: %s \n", err)
+		panic(err)
 	}
-
+	router.Use(middlewares.GinBodyLogMiddleware)
+	router.Use(errorsController.Handler)
+	routes.ApplicationV1Router(router, DB)
+	startServer(router)
 
 }
 
 func startServer(router http.Handler) {
 	viper.SetConfigFile("config.json")
 	if err := viper.ReadInConfig(); err != nil {
-		panic(fmt.Errorf("fatal error in config file: %s \n", err))
+		_ = fmt.Errorf("fatal error in config file: %s \n", err.Error())
+		panic(err)
+
 	}
 	serverPort := fmt.Sprintf(":%s", viper.GetString("ServerPort"))
 	s := &http.Server{
@@ -54,6 +50,8 @@ func startServer(router http.Handler) {
 		MaxHeaderBytes: 1 << 20,
 	}
 	if err := s.ListenAndServe(); err != nil {
-		panic(fmt.Errorf("fatal error description: %s \n", strings.ToLower(err.Error())))
+		_ = fmt.Errorf("fatal error description: %s \n", strings.ToLower(err.Error()))
+		panic(err)
+
 	}
 }
